move sketch attributes to a dedicated command group nested in sketch attribute group
